+++ public_api.ts
@@ -1,4 +1,4 @@
-import { InputFile } from './src/lib/dto/input-file';
+import { InputFile } from './src/lib/interfaces/input-file';
 import { InputFileComponent } from './src/lib/components/input-file/input-file.component';
 import { InputFileModule } from './src/lib/input-file.module';
 import { InputFileService } from './src/lib/services/input-file.service';